+++ app/Http/Controllers/Resume/ResumeController.php
@@ -10,6 +10,20 @@ use Inertia\Inertia;
 
 class ResumeController extends Controller
 {
+    public function index()
+    {
+        $resumes = Resume::get();
+
+        return Inertia::render('resumes/index', [
+            'resumes' => ResumeResource::collection($resumes),
+        ]);
+    }
+
+    public function create()
+    {
+        return Inertia::render('resumes/create');
+    }
+
     public function store(StoreResumeRequest $request)
     {
         $validated = $request->validated();
@@ -24,15 +38,15 @@ class ResumeController extends Controller
         });
     }
 
-    public function create()
-    {
-        return Inertia::render('resumes/create');
-    }
-
     public function edit(Resume $resume)
     {
         return Inertia::render('resumes/edit', [
             'resume' => new ResumeResource($resume),
         ]);
+    }
+
+    public function destroy(Resume $resume)
+    {
+        $resume->delete();
     }
 }